refactor: centralize shared API contract types
